drop openapi-genrator
I am having some issues with openapi-generator, AFIAIK it can't generate
proper code when requestBody is used in the specification. it was also
adding unecessary complexity to the code and slowing down the development
progress. instead of using openapi-generator I will just write a dead
simple connexion web server. maybe in the future we could go back to
openapi-generator but for now I just want to get the CI up and running.

Signed-off-by: Filipe Laíns <lains@archlinux.org>