device: add event_data fixture
Right now, if an error is thrown while we are sending the events, the
thread that is listening is  not be properly handled. This causes the
device to be destroyed with the listening thread still active, resulting
in the error below.

This patch introduced a new fixture, event_data, which will start
listening to events when the test starts.

```
Exception in thread Thread-1:
Traceback (most recent call last):
  File "/usr/lib/python3.8/threading.py", line 932, in _bootstrap_inner
    self.run()
  File "/usr/lib/python3.8/threading.py", line 870, in run
    self._target(*self._args, **self._kwargs)
  File "/home/anubis/git/ratbag-emu/tests/__init__.py", line 77, in collect_events
    evs += list(device.events())
  File "/usr/lib/python3.8/site-packages/libevdev/device.py", line 523, in events
    ev = self._libevdev.next_event(flags)
  File "/usr/lib/python3.8/site-packages/libevdev/_clib.py", line 872, in next_event
    raise OSError(-rc, os.strerror(-rc))
OSError: [Errno 19] No such device
```

Signed-off-by: Filipe Laíns <lains@archlinux.org>